+++ package.json
@@ -1,62 +1,44 @@
 {
-  "name": "material-ui-search-bar",
+  "name": "mui-search-bar",
   "version": "1.0.0",
-  "description": "Material style search bar",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "description": "Material UI style search bar refactored from https://www.npmjs.com/package/material-ui-search-bar using typescript and tested with jest",
+  "main": "dist/index.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "babel src -d lib --copy-files",
-    "prepare": "babel src -d lib --copy-files",
-    "styleguide": "styleguidist server",
-    "styleguide:build": "styleguidist build",
-    "publish-docs": "npm run styleguide:build && npm run gh-pages",
-    "gh-pages": "./node_modules/.bin/gh-pages -d styleguide",
-    "prettier": "prettier --write \"**/*.{json,md,js}\""
+    "build": "tsc"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/TeamWertarbyte/material-ui-search-bar.git"
+    "url": "git+https://github.com/SnekCode/material-ui-search-bar.git"
   },
   "keywords": [
     "react",
     "material",
     "material-design",
     "material-ui",
+    "mui",
     "search",
-    "react-component"
+    "react-component",
+    "tested",
+    "typescript"
   ],
-  "author": "Wertarbyte <kontakt@wertarbyte.com> (https://wertarbyte.com)",
+  "author": "SnekCode",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/TeamWertarbyte/material-ui-search-bar/issues"
+    "url": "https://github.com/SnekCode/material-ui-search-bar/issues"
   },
-  "homepage": "https://github.com/TeamWertarbyte/material-ui-search-bar#readme",
+  "homepage": "https://github.com/SnekCode/material-ui-search-bar#readme",
   "devDependencies": {
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.11.0",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "gh-pages": "^1.2.0",
-    "prettier": "^2.0.5",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
-    "react-styleguidist": "^11.0.8",
-    "webpack": "^4.44.1",
-    "webpack-blocks": "^2.1.0"
+    "typescript": "^3.9.7"
   },
   "peerDependencies": {
     "@material-ui/core": "^4.0.0",
     "@material-ui/icons": "^4.0.0",
-    "react": "^16.8.0"
+    "react": "^16.8.0",
+    "react-dom": "^16.10.2"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "prop-types": "^15.5.8"
   },
-  "standard": {
-    "parser": "babel-eslint"
-  }
 }
